third_party/blink/web_tests/fast/scroll-behavior/overflow-scroll-animates.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<title>Tests that overflow scrolls are animated.</title>

<div id="container" style="overflow: scroll;width: 200px;height: 200px;">
  <div id="content" style="width: 750px;height: 10000px;background-color: blue;"></div>
</div>

<script>
if (!window.eventSender || !window.internals || !window.testRunner) {
  assert_unreached("This test must be run in the test harness");
}
internals.settings.setScrollAnimatorEnabled(true);
testRunner.waitUntilDone();
testRunner.dumpAsText();

const container = document.getElementById("container");

let prevScrollTop;
container.addEventListener("scroll", function onElementScroll() {
  if (prevScrollTop === undefined) {
    assert_not_equals(container.scrollTop,container.scrollHeight - container.clientHeight);
    prevScrollTop = container.scrollTop;
  } else {
    assert_true(container.scrollTop > prevScrollTop);
    console.log('PASS');
    testRunner.notifyDone();
  }
});

eventSender.mouseMoveTo(100, 100);
eventSender.mouseDown();
eventSender.mouseUp();
eventSender.keyDown('End');
</script>
